+++ src/app/PubSubVitessceGrid.js
@@ -52,7 +52,7 @@ export default function PubSubVitessceGrid(props) {
   const containerRef = useRef();
 
   const margin = 10;
-  const padding = 10;
+  const padding = 5;
 
   // Detect when the `config` or `containerHeight` variables
   // have changed, and update `rowHeight` in response.